fixed and refactored
