+++ src/db.ts
@@ -1,7 +1,9 @@
 'use strict';
 
 import dotenv from 'dotenv';
-import { DataTypes, Sequelize } from 'sequelize';
+import { Sequelize } from 'sequelize';
+import { products } from './models/products.js';
+import { productsDetails } from './models/products-details.js';
 dotenv.config();
 
 const { DB_URL } = process.env;
@@ -21,146 +23,14 @@ export const sequelize = new Sequelize(DB_URL, {
   }
 });
 
-export const Products = sequelize.define(
-  'Products',
-  {
-    id: {
-      allowNull: false,
-      autoIncrement: true,
-      primaryKey: true,
-      type: DataTypes.INTEGER
-    },
-    category: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    itemId: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true
-    },
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    fullPrice: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-    price: {
-      type: DataTypes.DECIMAL
-    },
-    screen: {
-      type: DataTypes.STRING
-    },
-    capacity: {
-      type: DataTypes.STRING
-    },
-    color: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    ram: {
-      type: DataTypes.STRING
-    },
-    year: {
-      type: DataTypes.INTEGER
-    },
-    image: {
-      type: DataTypes.TEXT,
-      allowNull: false
-    }
-  },
-  {
-    tableName: 'Products',
-    createdAt: false,
-    updatedAt: false
-  }
-);
+export const Products = sequelize.define('Products', products, {
+  tableName: 'Products',
+  createdAt: false,
+  updatedAt: false
+});
 export const ProductsDetails = sequelize.define(
   'Products-details',
-  {
-    id: {
-      allowNull: false,
-      autoIncrement: true,
-      primaryKey: true,
-      type: DataTypes.STRING
-    },
-
-    namespaceId: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-
-    capacityAvailable: {
-      type: DataTypes.ARRAY(DataTypes.STRING)
-    },
-
-    capacity: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-
-    priceRegular: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-
-    priceDiscount: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
-    },
-
-    colorsAvailable: {
-      type: DataTypes.ARRAY(DataTypes.STRING)
-    },
-
-    color: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-
-    images: {
-      type: DataTypes.ARRAY(DataTypes.TEXT)
-    },
-
-    description: {
-      type: DataTypes.JSONB
-    },
-
-    screen: {
-      type: DataTypes.STRING
-    },
-
-    resolution: {
-      type: DataTypes.STRING
-    },
-
-    processor: {
-      type: DataTypes.STRING
-    },
-
-    ram: {
-      type: DataTypes.STRING
-    },
-
-    camera: {
-      type: DataTypes.STRING
-    },
-
-    zoom: {
-      type: DataTypes.STRING
-    },
-
-    cell: {
-      type: DataTypes.ARRAY(DataTypes.STRING)
-    }
-  },
+  productsDetails,
   {
     tableName: 'Products-details',
     createdAt: false,
